refactor(frontend): migrate shared components to components/, lib/, hooks/
Move generic utility modules from features/shared/ to canonical locations:

- components/ (ComingSoon, NotFound, PageState, TableRowActionsMenu)
- lib/ (constants, formatters, selectOnFocus)
- hooks/ (use-mobile)
- features/weddings/utils/ (weddingStatus)

Update all import paths across the codebase to match the new locations.
Remove stale CSS variables from index.css.
