+++ frontend/src/features/weddings/components/WeddingFilters.tsx
@@ -7,7 +7,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Search } from "lucide-react";
-import type { WeddingStatusFilter } from "@/features/shared/utils/weddingStatus";
+import type { WeddingStatusFilter } from "@/features/weddings/utils/wedding-status";
 
 interface WeddingFiltersProps {
   search: string;
