+++ frontend/src/index.css
@@ -15,8 +15,6 @@
     /* --- PALETA SIM, ACEITO! (Refinada - Violet Eye-Friendly) --- */
     --primary: 262 83% 58%;          /* #7C3AED - Violet 600 */
     --primary-foreground: 210 20% 98%; /* Quase branco */
-    --primary-hover: 262 80% 50%;    /* Violet 700 approx */
-    --primary-dark: 263 70% 22%;     /* Violet 950 */
 
     --secondary: 262 50% 98%;        /* Fundo bem leve arroxeado */
     --secondary-foreground: 262 70% 22%;
@@ -29,19 +27,15 @@
 
     --destructive: 340 82% 52%;      /* #E91E63 - Danger Base */
     --destructive-foreground: 340 100% 97%; /* #FCE4EC - Fundo Suave */
-    --destructive-dark: 340 77% 43%;  /* #C2185B - Texto Escuro */
 
     --success: 174 100% 29%;         /* #009688 - Success Base */
     --success-foreground: 174 33% 92%; /* #E0F2F1 - Fundo Suave */
-    --success-dark: 174 100% 21%;    /* #00796B - Texto Escuro */
 
     --warning: 45 100% 51%;          /* #FFC107 - Warning Base */
     --warning-foreground: 45 100% 94%;  /* #FFF8E1 - Fundo Suave */
-    --warning-dark: 38 100% 50%;     /* #FFA000 - Texto Escuro */
 
     --info: 199 98% 48%;             /* #03A9F4 - Info Base */
     --info-foreground: 199 100% 94%;    /* #E1F5FE - Fundo Suave */
-    --info-dark: 201 98% 41%;        /* #0288D1 - Texto Escuro */
 
     --border: 240 5.9% 90%;          /* Zinc 200 approx */
     --input: 240 5.9% 90%;
@@ -70,8 +64,6 @@
     /* --- INVERSÃO PROPORCIONAL SIM, ACEITO! (Violet) --- */
     --primary: 262 80% 70%;
     --primary-foreground: 263 70% 22%;
-    --primary-hover: 262 80% 80%;
-    --primary-dark: 210 20% 98%;
 
     --secondary: 262 30% 12%;
     --secondary-foreground: 210 20% 98%;
@@ -84,19 +76,15 @@
 
     --destructive: 340 75% 60%;
     --destructive-foreground: 340 77% 43%;
-    --destructive-dark: 340 100% 97%;
 
     --success: 174 65% 50%;
     --success-foreground: 174 100% 21%;
-    --success-dark: 174 33% 92%;
 
     --warning: 45 90% 60%;
     --warning-foreground: 38 100% 50%;
-    --warning-dark: 45 100% 94%;
 
     --info: 199 90% 60%;
     --info-foreground: 201 98% 41%;
-    --info-dark: 199 100% 94%;
 
     --border: 240 3.7% 15.9%;
     --input: 240 3.7% 15.9%;
@@ -123,8 +111,6 @@
   --color-popover-foreground: hsl(var(--popover-foreground));
   --color-primary: hsl(var(--primary));
   --color-primary-foreground: hsl(var(--primary-foreground));
-  --color-primary-hover: hsl(var(--primary-hover));
-  --color-primary-dark: hsl(var(--primary-dark));
 
   --color-secondary: hsl(var(--secondary));
   --color-secondary-foreground: hsl(var(--secondary-foreground));
@@ -135,19 +121,15 @@
 
   --color-destructive: hsl(var(--destructive));
   --color-destructive-foreground: hsl(var(--destructive-foreground));
-  --color-destructive-dark: hsl(var(--destructive-dark));
 
   --color-success: hsl(var(--success));
   --color-success-foreground: hsl(var(--success-foreground));
-  --color-success-dark: hsl(var(--success-dark));
 
   --color-warning: hsl(var(--warning));
   --color-warning-foreground: hsl(var(--warning-foreground));
-  --color-warning-dark: hsl(var(--warning-dark));
 
   --color-info: hsl(var(--info));
   --color-info-foreground: hsl(var(--info-foreground));
-  --color-info-dark: hsl(var(--info-dark));
 
   --color-border: hsl(var(--border));
   --color-input: hsl(var(--input));
@@ -167,27 +149,9 @@
   --radius-sm: calc(var(--radius) - 4px);
 
   /* --- DEFINIÇÕES DE ANIMAÇÃO --- */
-  --animate-accordion-down: accordion-down 0.2s ease-out;
-  --animate-accordion-up: accordion-up 0.2s ease-out;
   --animate-collapsible-down: collapsible-down 0.2s ease-out;
   --animate-collapsible-up: collapsible-up 0.2s ease-out;
 
-  @keyframes accordion-down {
-    from {
-      height: 0;
-    }
-    to {
-      height: var(--radix-accordion-content-height);
-    }
-  }
-  @keyframes accordion-up {
-    from {
-      height: var(--radix-accordion-content-height);
-    }
-    to {
-      height: 0;
-    }
-  }
   @keyframes collapsible-down {
     from {
       height: 0;
